requirements.txt
unstructured[all-docs]
unstructured-client
python-dotenv
langchain
langchain_core
langchain_community
streamlit
fpdf